Cleaning up

+++ xflr5/xflr5-gui/viewwidgets/section2dwidget.cpp
@@ -1065,9 +1065,9 @@ void Section2dWidget::drawBackImage(QPainter &painter)
         int w = (int)((double)m_BackImage.width()* xscale);
         int h = (int)((double)m_BackImage.height()* yscale);
 		//the coordinates of the top left corner are measured from the center of the viewport
-		double xtop = VCenter.x() + m_ViewportTrans.x() - (int)((double)m_BackImage.width()  /2.*xscale);
-		double ytop = VCenter.y() + m_ViewportTrans.y() - (int)((double)m_BackImage.height() /2.*yscale);
-//qDebug(" %11.5f   %11.5f", m_ViewportTrans.x()-m_ptOffset.x(), m_ViewportTrans.y()-m_ptOffset.y());
+        double xtop = VCenter.x() + m_ViewportTrans.x() - (double)m_BackImage.width()  /2.*xscale;
+        double ytop = VCenter.y() + m_ViewportTrans.y() - (double)m_BackImage.height() /2.*yscale;
+
         painter.drawPixmap(xtop, ytop, w, h, m_BackImage);
 		painter.restore();
     }
